+++ docs/src/lib.rs
@@ -1,0 +1,5 @@
+#![feature(lazy_cell)]
+
+pub mod page;
+pub mod string_writer;
+pub mod views;